src/requirements.txt
azure-storage-blob
azure-servicebus
azure-identity
azure-keyvault